+++ src/oph_job_list.h
@@ -50,6 +50,7 @@ typedef struct _oph_job_info {
 typedef struct {
 	oph_job_info *head;
 	oph_job_info *tail;
+	oph_job_info *saved;
 	unsigned int counter;
 } oph_job_list;
 
@@ -59,6 +60,8 @@ oph_job_info *oph_find_job_in_job_list(oph_job_list * list, int jobid, oph_job_i
 int oph_insert_into_job_list(oph_job_list * list, oph_job_info * item);
 int oph_drop_from_job_list(oph_job_list * list, oph_job_info * item, oph_job_info * prev);
 int oph_delete_from_job_list(oph_job_list * list, oph_job_info * item, oph_job_info * prev);
+int oph_save_job_in_job_list(oph_job_list * list, oph_job_info * item);
+int oph_delete_saved_jobs_from_job_list(oph_job_list * list);
 
 int oph_free_job_list(oph_job_list * list);